+++ easyarmorstands-api/src/main/java/me/m56738/easyarmorstands/api/element/ElementTypeRegistry.java
@@ -6,7 +6,7 @@ import org.jspecify.annotations.Nullable;
 
 @NullMarked
 public interface ElementTypeRegistry {
-    void register(Key key, ElementType type);
+    void register(ElementType type);
 
     @Nullable ElementType getOrNull(Key key);