+++ easyarmorstands-plugin/src/main/java/me/m56738/easyarmorstands/element/EntityElementKeys.java
@@ -1,0 +1,11 @@
+package me.m56738.easyarmorstands.element;
+
+import me.m56738.easyarmorstands.api.EasyArmorStands;
+import org.bukkit.NamespacedKey;
+
+public final class EntityElementKeys {
+    private EntityElementKeys() {
+    }
+
+    public static final NamespacedKey ELEMENT_TYPE = EasyArmorStands.key("element_type");
+}