+++ easyarmorstands-plugin/src/main/java/me/m56738/easyarmorstands/element/SimpleEntityElementProvider.java
@@ -3,22 +3,38 @@ package me.m56738.easyarmorstands.element;
 import me.m56738.easyarmorstands.EasyArmorStandsPlugin;
 import me.m56738.easyarmorstands.api.element.Element;
 import me.m56738.easyarmorstands.api.element.EntityElementProvider;
-import me.m56738.easyarmorstands.util.Util;
+import me.m56738.easyarmorstands.api.element.EntityElementType;
+import net.kyori.adventure.key.Key;
 import org.bukkit.entity.Entity;
-import org.bukkit.entity.Player;
-import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+import org.jspecify.annotations.NullMarked;
 
-public class SimpleEntityElementProvider implements EntityElementProvider {
-    @Override
-    public Element getElement(@NotNull Entity entity) {
-        if (entity instanceof Player && !entity.hasMetadata("NPC") && !EasyArmorStandsPlugin.getInstance().getConfiguration().editor.allowPlayers) {
-            return null;
-        }
-        return new SimpleEntityElementType<>(entity.getType(), Util.getEntityClass(entity)).getElement(entity);
+@NullMarked
+public class SimpleEntityElementProvider<E extends Entity> implements EntityElementProvider {
+    private final EntityElementType<E> type;
+
+    public SimpleEntityElementProvider(EntityElementType<E> type) {
+        this.type = type;
     }
 
     @Override
-    public @NotNull Priority getPriority() {
-        return Priority.LOWEST;
+    public Key key() {
+        return type.key();
+    }
+
+    @Override
+    public boolean isApplicable(Entity entity) {
+        if (!EasyArmorStandsPlugin.getInstance().getConfiguration().editor.allowEntities) {
+            return false;
+        }
+        return type.getEntityClass().isInstance(entity);
+    }
+
+    @Override
+    public @Nullable Element getElement(Entity entity) {
+        if (type.getEntityClass().isInstance(entity)) {
+            return type.getElement(type.getEntityClass().cast(entity));
+        }
+        return null;
     }
 }